Added Attack Droid
Update script names to fit additional enemy types
Added model for attack droid
Working on new scripts for different droids and spawning
