--- Assets/Gurdian Outpost Bots.blend1.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 75dabab6ea0cc074791ec06284073539
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
