--- Assets/Lightsaber.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: ad329735e3c360949bd3396f43b392a8
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
