--- Jedi Trainer VR/Assets/LaserbeamController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 98c3dc53bb3dc0047a02d2d19ab4588a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
