+++ Jedi Trainer VR/Assets/Prefabs/TrainingDroid.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 38da0997d8e38d048a9c5b2e80d64835
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
