--- Jedi Trainer VR/Assets/Scripts/EnemyController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cb9bcf4bcc4f3f3439ad678ff2524f67
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
