--- Jedi Trainer VR/Assets/Scripts/EnemyMovement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 146603ed2b4ad884c853518a74086da1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
